pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "react-agent"
version = "0.1.0"
description = "Multi-tool ReAct agent (calculator, web search, memory)"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.40.0",
  "duckduckgo-search>=6.0.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
react-agent = "react_agent.main:main"

[tool.setuptools.packages.find]
where = ["src"]
